Add build-step setting the folding to the T^2 cycles per sequence target
This is a more robust solution compared to the previous, manual folding
configuration, as this relies on the existing SetFolding transformation
from FINN, which already handles MVAU and Thresholding layers.

This enables using RTL varaints of the MVAU and Thresholding layers, as
issues due to manual folding preventing this are avoided.